Use anvil on dagger

+++ domain/src/main/java/akio/apps/myrun/domain/strava/DeauthorizeStravaUsecase.kt
@@ -2,11 +2,13 @@ package akio.apps.myrun.domain.strava
 
 import akio.apps.myrun.data.authentication.api.UserAuthenticationState
 import akio.apps.myrun.data.eapps.api.ExternalAppProvidersRepository
+import akio.apps.myrun.data.eapps.api.StravaSyncState
 import akio.apps.myrun.data.eapps.api.StravaTokenRepository
 import javax.inject.Inject
 
 class DeauthorizeStravaUsecase @Inject constructor(
     private val stravaTokenRepository: StravaTokenRepository,
+    private val stravaSyncState: StravaSyncState,
     private val externalAppProvidersRepository: ExternalAppProvidersRepository,
     private val userAuthenticationState: UserAuthenticationState
 ) {
@@ -15,9 +17,10 @@ class DeauthorizeStravaUsecase @Inject constructor(
             ?: return
 
         val stravaToken = externalAppProvidersRepository.getStravaProviderToken(userAccountId)
-            ?: return
-
-        stravaTokenRepository.deauthorizeToken(stravaToken.accessToken)
-        externalAppProvidersRepository.removeStravaProvider(userAccountId)
+        if (stravaToken != null) {
+            stravaTokenRepository.deauthorizeToken(stravaToken.accessToken)
+            externalAppProvidersRepository.removeStravaProvider(userAccountId)
+        }
+        stravaSyncState.setStravaSyncAccountId(null)
     }
 }
